requirements.txt
absl-py==2.1.0
ase==3.23.0
astunparse==1.6.3
axial_positional_embedding==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.25.0
CoLT5-attention==0.11.0
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
lit==15.0.7
local-attention==1.9.14
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
product_key_memory==0.2.11
protobuf==4.25.4
psutil==6.0.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
reformer-pytorch==1.4.4
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
sentry-sdk==2.12.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.5
triton==2.0.0
typing_extensions==4.9.0
urllib3==2.2.2
wandb==0.17.6
Werkzeug==3.0.3
wrapt==1.16.0